+++ changelog.md
@@ -13,6 +13,7 @@
 - Added use of right-clicking mouse as alternative for drawing
   (can be handy for left-handed people!)
 - Brush size was changed to 15 (previously: 1), so it doesn't need scaling up instantly
+- Made brush switch slower, so it won't skip brushes
 
 ### 0.1.0
 - Added basic colours